requirements/dev.txt
-r _base.txt
django-debug-toolbar==1.10.1